Taşınabilirlik: Statfs alan dönüşümleri — panel artık FreeBSD'ye derleniyor
BSD kararını kayda geçirirken kanıt derlemesi tek pürüzü yakaladı:
Statfs_t alan tipleri OS'lere göre değişiyor (Linux uint64, FreeBSD
int64). Açık uint64 dönüşümüyle panel ve agent artık GOOS=freebsd ile
derleniyor — yol haritasındaki 'seçenek korunur' iddiası somut kanıtlı.
Linux davranışı değişmedi.

+++ cmd/panel/system_stats.go
@@ -148,9 +148,15 @@ func readDisk(path string) (used, total uint64) {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return 0, 0
 	}
+	// Explicit casts: Statfs_t field types differ across OSes (Linux uses
+	// uint64, FreeBSD int64/uint32 mixes) — this line is what keeps the
+	// panel cross-compiling to BSD, per the roadmap's preserved option.
+	// Açık dönüşümler: Statfs_t alan tipleri OS'lere göre değişir (Linux
+	// uint64, FreeBSD int64/uint32 karışık) — panelin BSD'ye çapraz
+	// derlenmesini koruyan satır budur.
 	bs := uint64(st.Bsize)
-	total = st.Blocks * bs
-	free := st.Bavail * bs
+	total = uint64(st.Blocks) * bs
+	free := uint64(st.Bavail) * bs
 	if free > total {
 		free = total
 	}
